backend/requirements.txt
fastapi
uvicorn
scikit-learn
numpy
matplotlib
pillow
python-multipart
